# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Blacket-style game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get recent chat messages (newest last)
      parameters:
        - name: after
          in: query
          required: false
          schema:
            type: integer
          description: Only return messages with id greater than this
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
        "400":
          description: Invalid message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "403":
          description: You are muted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /chat/messages/{id}:
    delete:
      operationId: deleteChatMessage
      tags: [chat]
      summary: Delete your own chat message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Message not found or not yours
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    ChatMessage:
      type: object
      properties:
        id: { type: integer }
        author: { type: string }
        avatarBlook: { type: ["string", "null"] }
        avatarImage: { type: ["string", "null"] }
        nameEffect: { type: ["string", "null"] }
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
        content: { type: string }
        mentions:
          type: array
          items: { type: string }
        mentionEffects:
          type: object
          additionalProperties: { type: ["string", "null"] }
        isMine: { type: boolean }
        createdAt: { type: string }
        clanName: { type: ["string", "null"] }
        clanColor: { type: ["string", "null"] }
        chatColor: { type: ["string", "null"] }
      required: [id, author, avatarBlook, avatarImage, badges, content, mentions, isMine, createdAt, clanName, clanColor]
    ChatMessageInput:
      type: object
      properties:
        content: { type: string, minLength: 1, maxLength: 500 }
      required: [content]
...
    Badge:
      type: object
      properties:
        name: { type: string }
        image: { type: string }
        description: { type: string }
      required: [name, image, description]
...